+++ site/src/docs.config.ts
@@ -2,7 +2,8 @@ import type { Sidebar } from '@/types/docs';
 
 export const sidebar: Sidebar = [
   {
-    sidebarLabel: 'Getting started',
+    sidebarLabel: 'Site development',
+    devOnly: true,
     contents: [
       {
         slug: 'how-to/write-guides',
@@ -14,18 +15,25 @@ export const sidebar: Sidebar = [
         sidebarLabel: 'Writing references',
         devOnly: true,
       },
+    ],
+  },
+  {
+    sidebarLabel: 'Getting started',
+    contents: [
       { slug: 'how-to/installation' },
+      { slug: 'concepts/overview' },
       { slug: 'concepts/v10-roadmap', sidebarLabel: 'Roadmap' },
     ],
   },
-  // {
-  //   sidebarLabel: 'Concepts',
-  //   contents: [
-  //     { slug: 'concepts/architecture' },
-  //     { slug: 'concepts/skins' },
-  //     { slug: 'concepts/ui-components' }
-  //   ],
-  // },
+  {
+    sidebarLabel: 'Concepts',
+    contents: [
+      { slug: 'concepts/features' },
+      { slug: 'concepts/skins' },
+      { slug: 'concepts/presets' },
+      { slug: 'concepts/ui-components' },
+    ],
+  },
   {
     sidebarLabel: 'How to',
     contents: [{ slug: 'how-to/customize-skins' }],
@@ -33,6 +41,8 @@ export const sidebar: Sidebar = [
   {
     sidebarLabel: 'Components',
     contents: [
+      { slug: 'reference/player-provider' },
+      { slug: 'reference/player-container' },
       // sorted alphabetically
       { slug: 'reference/buffering-indicator' },
       { slug: 'reference/captions-button' },
@@ -45,9 +55,11 @@ export const sidebar: Sidebar = [
       { slug: 'reference/popover' },
       { slug: 'reference/poster' },
       { slug: 'reference/seek-button' },
+      { slug: 'reference/slider' },
       { slug: 'reference/thumbnail' },
       { slug: 'reference/time' },
       { slug: 'reference/time-slider' },
+      { slug: 'reference/tooltip' },
       { slug: 'reference/volume-slider' },
     ],
   },
@@ -79,7 +91,6 @@ export const sidebar: Sidebar = [
     contents: [
       { slug: 'reference/html-create-player', sidebarLabel: 'createPlayer' },
       { slug: 'reference/player-controller' },
-      { slug: 'reference/player-mixin', sidebarLabel: 'PlayerMixin (removed)' },
       {
         sidebarLabel: 'Advanced',
         defaultOpen: false,
@@ -94,18 +105,20 @@ export const sidebar: Sidebar = [
     ],
   },
   {
-    sidebarLabel: 'Selectors',
+    sidebarLabel: 'Features',
     contents: [
       { slug: 'reference/create-selector' },
-      { slug: 'reference/select-buffer' },
-      { slug: 'reference/select-controls' },
-      { slug: 'reference/select-fullscreen' },
-      { slug: 'reference/select-pip', sidebarLabel: 'selectPiP' },
-      { slug: 'reference/select-playback' },
-      { slug: 'reference/select-playback-rate' },
-      { slug: 'reference/select-source' },
-      { slug: 'reference/select-time' },
-      { slug: 'reference/select-volume' },
+      { slug: 'reference/feature-buffer' },
+      { slug: 'reference/feature-controls' },
+      { slug: 'reference/feature-error' },
+      { slug: 'reference/feature-fullscreen' },
+      { slug: 'reference/feature-pip', sidebarLabel: 'Picture-in-picture' },
+      { slug: 'reference/feature-playback' },
+      { slug: 'reference/feature-playback-rate' },
+      { slug: 'reference/feature-source' },
+      { slug: 'reference/feature-text-tracks' },
+      { slug: 'reference/feature-time' },
+      { slug: 'reference/feature-volume' },
     ],
   },
 ];